FF-2013: add flow wizard
