+++ Client/Components/Common/FlowRadioGroup/FlowRadioGroup.razor
@@ -10,8 +10,8 @@
             @ondblclick="() => DblClick(item.Value)">
             <i class="@item.Icon"></i>
             <div class="info">
-                <span>@item.Title</span>
-                <p>@item.Description</p>
+                <span>@item.RenderTitle</span>
+                <p>@item.RenderDescription</p>
             </div>
             @if (item.ChildContent != null)
             {
