+++ Client/Wizards/NewLibraryWizard/NewLibraryWizard.razor
@@ -1,4 +1,4 @@
-@namespace FileFlows.Client.Components.Dialogs.Wizards
+@namespace FileFlows.Client.Wizards
 
 @using InputFile = FileFlows.Client.Components.Inputs.InputFile
 @using InputText = FileFlows.Client.Components.Inputs.InputText
