+++ Shared/Models/FlowTemplate.cs
@@ -173,4 +173,14 @@ public class FlowListModel: IInUse, IUniqueObject<Guid>
     /// Gets or sets if this flow is read only
     /// </summary>
     public bool ReadOnly { get; set; }
+    
+    /// <summary>
+    /// Gets or sets the description of the flow
+    /// </summary>
+    public string Description { get; set; }
+    
+    /// <summary>
+    /// Gets or sets the Icon of the flow
+    /// </summary>
+    public string Icon { get; set; }
 }
